+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 14.0.2
 
+* [THORN-322] Add docs for Health of hardware facing devices.
 * [SKB-1445] Update ``AcquireDataForCalibration`` to be more resilient to DAQ errors.
 * [THORN-125] Improve observability when UNPROGRAMMED.
 * [THORN-125] Bump daq version: 6.1.0 -> 6.1.1